vc12

+++ app/api/image/delete/route.ts
@@ -1,22 +1,36 @@
 import driveService from "@/configs/google";
-import { NextResponse } from "next/server";
+import { getToken } from "next-auth/jwt";
+import { NextRequest, NextResponse } from "next/server";
 
-export async function POST(req: Request) {
-    try {
-        const { id } = (await req.json()) as { id: string };
-        if (!id) {
-            return NextResponse.json("");
+export async function POST(req: NextRequest) {
+    const token = await getToken({ req });
+
+    if (token && token.role === "admin") {
+        try {
+            const { id } = (await req.json()) as { id: string };
+            if (!id) {
+                return NextResponse.json("");
+            }
+            const res = await driveService.files.delete({
+                fileId: id,
+            });
+            return NextResponse.json(res);
+        } catch (err: any) {
+            return NextResponse.json({
+                status: err.status,
+                errors: err.errors.map((error: any) => {
+                    return { message: error.message };
+                }),
+            });
         }
-        const res = await driveService.files.delete({
-            fileId: id,
-        });
-        return NextResponse.json(res);
-    } catch (err: any) {
-        return NextResponse.json({
-            status: err.status,
-            errors: err.errors.map((error: any) => {
-                return { message: error.message };
-            }),
-        });
     }
+
+    return NextResponse.json({
+        status: 1020,
+        errors: [
+            {
+                message: "Отказано в доступе",
+            },
+        ],
+    });
 }
